requirements/base.txt
Django==4.1.7
psycopg2-binary==2.9.5
whitenoise==6.4.0
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2